feat(code-viewer): remove Suspense
